feat: Implement turn-based action system and player movement
This commit introduces a turn-based action system, allowing both the player and enemies to take turns.

Key changes:

-   **Turn System:**
    -   Implemented a `TurnSystem` resource to manage turn order and timing.
    -   Added `TurnActor` component to entities to define their speed and next turn time.
    -   `process_turns` system now determines whose turn it is and handles player input or AI actions.
    -   `execute_actions` system processes actions like movement.
    -   `TurnQueue` resource has been replaced by `TurnSystem`
-   **Player Input:**
    -   `player_input_system` now handles keyboard input, mapping keys to actions.
    -   Player actions (move, wait) are now processed and scheduled.
    -   Added `WaitingForInput` component to pause turn processing when the player's turn arrives.
-   **Enemy AI:**
    -   Basic enemy AI implemented to move in a random direction.
-   **Movement:**
    -   `TryMove` command now handles movement logic, checking for collisions with walls.
    -   `MoveDirection` enum added to represent movement directions.
    -   `Position` component now supports addition with `MoveDirection`.
-   **Spawning:**
    -   `spawn_player` now spawns both the player and an enemy.
    -   Player and enemy are placed on valid floor tiles.
- **Rendering**
    - Added a basic text based rendering system for tiles.
-   **Code Structure:**
    -   Refactored code into `model`, `controller`, and `view` modules.
    -   Improved code organization and readability.
- **Other**
    - Added a `.gitignore` file to the `brtk` folder.

This commit lays the foundation for a more complex turn-based game with player and enemy interactions.

+++ src/model/model_constants.rs
@@ -1,6 +1,6 @@
 pub struct ModelConstants;
 
 impl ModelConstants {
-    pub const MAP_WIDTH: u32 = 40;
-    pub const MAP_HEIGHT: u32 = 30;
+    pub const MAP_WIDTH: usize = 40;
+    pub const MAP_HEIGHT: usize = 30;
 }
